requirements.txt
langgraph>=0.4.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
langfuse>=2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
python-multipart>=0.0.12
pydantic>=2.10.0
pydantic-settings>=2.6.0
asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0.0
redis>=5.0.0
docker>=7.0.0
httpx>=0.28.0
pyyaml>=6.0.0
aiofiles>=24.0.0
